+++ packages/atomic/cypress/integration/result-list/result-components/result-text.cypress.ts
@@ -126,6 +126,8 @@ describe('Result Text Component', () => {
         ResultTextSelectors.firstInResult().should('have.text', rawValue);
         ResultTextSelectors.highlight().should('have.text', highlightedValue);
       });
+
+      CommonAssertions.assertAccessibility(ResultTextSelectors.firstInResult);
     });
 
     describe('when the "shouldHighlight" prop is false', () => {