+++ packages/atomic/cypress/integration/result-list/result-list.cypress.ts
@@ -1,8 +1,12 @@
 import {createAliasNavigation, PagerSelectors} from '../pager-selectors';
 import {withAnySectionnableResultList} from './result-list-utils';
-import {ResultListSelectors} from './result-list-selectors';
+import {
+  resultListComponent,
+  ResultListSelectors,
+} from './result-list-selectors';
 import {generateComponentHTML, TestFixture} from '../../fixtures/test-fixture';
 import {addResultList, buildTemplateWithSections} from './result-list-actions';
+import * as CommonAssertions from '../common-assertions';
 
 describe('Result List Component', () => {
   it('should load', () => {
@@ -145,10 +149,11 @@ describe('Result List Component', () => {
         .init();
     });
 
-    withAnySectionnableResultList(() => {
-      it.skip('should pass accessibility tests', () => {
-        cy.checkA11y();
-      });
-    });
+    withAnySectionnableResultList(
+      () => {
+        CommonAssertions.assertAccessibility(resultListComponent);
+      },
+      {densities: ['normal'], imageSizes: ['icon', 'small']}
+    );
   });
 });